Refactor WhiteBalanceV2.kt

+++ lightcrafts/src/com/lightcrafts/extensions/primitive/Float.kt
@@ -5,4 +5,4 @@ package com.lightcrafts.extensions.primitive
 import kotlin.math.max
 import kotlin.math.min
 
-fun Double.clamp(lo: Double, hi: Double): Double = max(lo, min(hi, this))
+fun Float.clamp(lo: Float, hi: Float): Float = max(lo, min(hi, this))
